feat: インターバルタイマーを実行中のセットに追従させる
セット数が増えると下方向にスクロールが必要になり、カード上部に固定
されていたタイマーが画面外に出て操作しづらくなっていた。

タイマーをアクティブなセット行の直下に表示し、フォーカス移動に追従
させる。実行中のセットが無いときはセット枠の先頭が定位置。

状態は useRestTimer フックへ切り出した。表示位置が変わると RestTimer
は unmount/remount されるため、状態を表示コンポーネント内に持ったまま
だとカウントダウンが毎回リセットされてしまう。

タイマーの開始・リセットは従来どおりユーザーのタップのみで、セット
フォーカスの操作とは意図的に連動させていない。
