+++ packages/frontend/src/components/exercise/RestTimer.tsx
@@ -1,136 +1,23 @@
-import { useState, useEffect, useRef, useCallback } from 'react';
+import type { RestTimerController } from '../../hooks/useRestTimer';
 
 interface RestTimerProps {
-  defaultSeconds?: number;
-  incrementSeconds?: number;
+  timer: RestTimerController;
 }
 
-export function RestTimer({ defaultSeconds = 60, incrementSeconds = 60 }: RestTimerProps) {
-  const [seconds, setSeconds] = useState(defaultSeconds);
-  const [isRunning, setIsRunning] = useState(false);
-  const [totalSeconds, setTotalSeconds] = useState(defaultSeconds);
-  const touchStartX = useRef<number | null>(null);
-  const audioContextRef = useRef<AudioContext | null>(null);
-
-  // Play alarm sound using Web Audio API
-  const playAlarm = useCallback(() => {
-    try {
-      if (!audioContextRef.current) {
-        audioContextRef.current = new AudioContext();
-      }
-      const ctx = audioContextRef.current;
-      const oscillator = ctx.createOscillator();
-      const gainNode = ctx.createGain();
-
-      oscillator.connect(gainNode);
-      gainNode.connect(ctx.destination);
-
-      oscillator.frequency.value = 800;
-      oscillator.type = 'sine';
-      gainNode.gain.value = 0.3;
-
-      // Beep pattern: 3 short beeps
-      const now = ctx.currentTime;
-      oscillator.start(now);
-
-      gainNode.gain.setValueAtTime(0.3, now);
-      gainNode.gain.setValueAtTime(0, now + 0.15);
-      gainNode.gain.setValueAtTime(0.3, now + 0.25);
-      gainNode.gain.setValueAtTime(0, now + 0.4);
-      gainNode.gain.setValueAtTime(0.3, now + 0.5);
-      gainNode.gain.setValueAtTime(0, now + 0.65);
-
-      oscillator.stop(now + 0.7);
-    } catch {
-      // Audio not supported or blocked
-    }
-  }, []);
-
-  // Timer countdown logic
-  useEffect(() => {
-    if (!isRunning) return;
-
-    const interval = setInterval(() => {
-      setSeconds((prev) => {
-        if (prev <= 1) {
-          setIsRunning(false);
-          playAlarm();
-          return 0;
-        }
-        return prev - 1;
-      });
-    }, 1000);
-
-    return () => clearInterval(interval);
-  }, [isRunning, playAlarm]);
-
-  // Handle tap: start or reset
-  const handleTap = () => {
-    if (isRunning) {
-      // Reset when tapped during countdown
-      setIsRunning(false);
-      setSeconds(totalSeconds);
-    } else {
-      // Start countdown
-      if (seconds === 0) {
-        setSeconds(totalSeconds);
-      }
-      setIsRunning(true);
-    }
-  };
-
-  // Handle swipe start
-  const handleTouchStart = (e: React.TouchEvent) => {
-    touchStartX.current = e.touches[0]?.clientX ?? null;
-  };
-
-  // Handle swipe end - right swipe adds time
-  const handleTouchEnd = (e: React.TouchEvent) => {
-    if (touchStartX.current === null) return;
-
-    const touchEndX = e.changedTouches[0]?.clientX ?? 0;
-    const diff = touchEndX - touchStartX.current;
-
-    // Right swipe (threshold: 50px)
-    if (diff > 50) {
-      const newTotal = totalSeconds + incrementSeconds;
-      setTotalSeconds(newTotal);
-      if (!isRunning) {
-        setSeconds(newTotal);
-      } else {
-        setSeconds((prev) => prev + incrementSeconds);
-      }
-    }
-    // Left swipe - decrease time (minimum 60 seconds)
-    else if (diff < -50 && totalSeconds > incrementSeconds) {
-      const newTotal = totalSeconds - incrementSeconds;
-      setTotalSeconds(newTotal);
-      if (!isRunning) {
-        setSeconds(newTotal);
-      } else {
-        setSeconds((prev) => Math.max(1, prev - incrementSeconds));
-      }
-    }
-
-    touchStartX.current = null;
-  };
-
-  // Format seconds as MM:SS
-  const formatTime = (secs: number) => {
-    const mins = Math.floor(secs / 60);
-    const remainingSecs = secs % 60;
-    return `${mins}:${remainingSecs.toString().padStart(2, '0')}`;
-  };
-
-  // Calculate progress percentage
-  const progress = totalSeconds > 0 ? (seconds / totalSeconds) * 100 : 0;
+/**
+ * インターバルタイマーの表示専用コンポーネント。
+ * 状態は useRestTimer が持つので、アクティブなセット行に追従して
+ * JSX上の位置が変わっても（＝remountされても）カウントダウンは継続する。
+ */
+export function RestTimer({ timer }: RestTimerProps) {
+  const { seconds, isRunning, progress, formattedTime, onTap, onTouchStart, onTouchEnd } = timer;
 
   return (
     <div
       className="select-none cursor-pointer"
-      onClick={handleTap}
-      onTouchStart={handleTouchStart}
-      onTouchEnd={handleTouchEnd}
+      onClick={onTap}
+      onTouchStart={onTouchStart}
+      onTouchEnd={onTouchEnd}
     >
       <div className="flex items-center gap-2">
         {/* Circular progress indicator */}
@@ -182,7 +69,7 @@ export function RestTimer({ defaultSeconds = 60, incrementSeconds = 60 }: RestTi
             isRunning ? 'text-orange-600' : seconds === 0 ? 'text-green-600' : 'text-gray-700'
           }`}
         >
-          {formatTime(seconds)}
+          {formattedTime}
         </span>
       </div>
     </div>
